Major cube_routes refactor
