+++ app.js
@@ -170,9 +170,9 @@ app.use('/cube/api/cubeJSON', apiLimiter);
 // Route files; they manage their own CSRF protection
 app.use('', require('./routes/root'));
 
-app.use('/cube', require('./routes/cube_routes'));
+app.use('/cube', require('./routes/cube/index'));
 app.use('/user', require('./routes/users_routes'));
-app.use('/dev', require('./routes/dev_routes'));
+app.use('/dev', require('./routes/cube/index'));
 app.use('/tool', require('./routes/tools_routes'));
 app.use('/comment', require('./routes/comment_routes'));
 app.use('/admin', require('./routes/admin_routes'));
