+++ src/pages/CommentPage.js
@@ -25,7 +25,7 @@ const translateType = {
 
 const translateLink = {
   comment: (id) => `/comment/${id}`,
-  blog: (id) => `/cube/blogpost/${id}`,
+  blog: (id) => `/cube/blog/blogpost/${id}`,
   deck: (id) => `/cube/deck/${id}`,
   card: (id) => `/tool/card/${id}`,
   article: (id) => `/content/article/${id}`,
