+++ src/pages/CubeDeckPage.js
@@ -84,7 +84,7 @@ const CubeDeckPage = ({ user, cube, deck, draft, loginCallback }) => {
             className="d-none"
             innerRef={submitDeckForm}
             method="POST"
-            action={`/cube/submitdeck/${cube._id}`}
+            action={`/cube/deck/submitdeck/${cube._id}`}
           >
             <Input type="hidden" name="body" value={draftId} />
             <Input type="hidden" name="skipDeckbuilder" value="true" />
@@ -120,7 +120,7 @@ const CubeDeckPage = ({ user, cube, deck, draft, loginCallback }) => {
                 </NavItem>
                 {user && deck.owner === user.id && (
                   <NavItem>
-                    <NavLink href={`/cube/deckbuilder/${deck._id}`}>Edit</NavLink>
+                    <NavLink href={`/cube/deck/deckbuilder/${deck._id}`}>Edit</NavLink>
                   </NavItem>
                 )}
                 {loading && <Spinner className="position-absolute" />}
@@ -130,14 +130,14 @@ const CubeDeckPage = ({ user, cube, deck, draft, loginCallback }) => {
                       Rebuild/Redraft
                     </DropdownToggle>
                     <DropdownMenu right>
-                      <DropdownItem href={`/cube/redraft/${deck._id}`}>Redraft</DropdownItem>
+                      <DropdownItem href={`/cube/deck/redraft/${deck._id}`}>Redraft</DropdownItem>
                       <DropdownItem onClick={haveBotsRedraft}>Have Bots Redraft</DropdownItem>
-                      <DropdownItem href={`/cube/rebuild/${deck._id}/${seatIndex}`}>Clone and Rebuild</DropdownItem>
+                      <DropdownItem href={`/cube/deck/rebuild/${deck._id}/${seatIndex}`}>Clone and Rebuild</DropdownItem>
                     </DropdownMenu>
                   </UncontrolledDropdown>
                 ) : (
                   <NavItem>
-                    <NavLink href={`/cube/rebuild/${deck._id}/${seatIndex}`}>Clone and Rebuild</NavLink>
+                    <NavLink href={`/cube/deck/rebuild/${deck._id}/${seatIndex}`}>Clone and Rebuild</NavLink>
                   </NavItem>
                 )}
                 <CustomImageToggler />
