+++ src/pages/CubeDecksPage.js
@@ -18,7 +18,7 @@ const CubeDecksPage = ({ user, cube, decks, pages, activePage, loginCallback }) 
     <DynamicFlash />
     <CubeLayout cube={cube} activeLink="playtest">
       <div className="my-3">
-        {pages > 1 && <Paginate count={pages} active={activePage} urlF={(i) => `/cube/decks/${cube._id}/${i}`} />}
+        {pages > 1 && <Paginate count={pages} active={activePage} urlF={(i) => `/cube/deck/decks/${cube._id}/${i}`} />}
         <Card>
           <CardHeader>
             <h5 className="mb-0">All Decks</h5>
@@ -29,12 +29,12 @@ const CubeDecksPage = ({ user, cube, decks, pages, activePage, loginCallback }) 
                 key={deck._id}
                 deck={deck}
                 canEdit={user?.id === deck.owner}
-                nextURL={`/cube/decks/${cube._id}/${activePage}`}
+                nextURL={`/cube/deck/decks/${cube._id}/${activePage}`}
               />
             ))}
           </CardBody>
         </Card>
-        {pages > 1 && <Paginate count={pages} active={activePage} urlF={(i) => `/cube/decks/${cube._id}/${i}`} />}
+        {pages > 1 && <Paginate count={pages} active={activePage} urlF={(i) => `/cube/deck/decks/${cube._id}/${i}`} />}
       </div>
     </CubeLayout>
   </MainLayout>
