+++ src/pages/GridDraftPage.js
@@ -356,7 +356,7 @@ const GridDraftPage = ({ user, cube, initialDraft, loginCallback }) => {
             className="d-none"
             innerRef={submitDeckForm}
             method="POST"
-            action={`/cube/submitgriddeck/${initialDraft.cube}`}
+            action={`/cube/deck/submitgriddeck/${initialDraft.cube}`}
           >
             <Input type="hidden" name="body" value={initialDraft._id} />
           </CSRFForm>
